qemu backend: honor DEFAULT_CPU and DEFAULT_RAM
When using qemu the upgrade test can work in two modes: by calling adt-run
(autopkgtest) with the qemu backend, or in "keep-overlay" mode, where a qemu
VM is started by the test harness and adt-run is made use it via its SSH
backend. The DEFAULT_CPU and DEFAULT_RAM settings were used only in
keep-overlay mode, and ignored when using the qemu backend directly.

This commit passes DEFAULT_CPU and DEFAULT_RAM to the qemu backend, using the
flags specified in autopkgtest-virt-qemu(1). 

The DEFAULT_RAM setting has been changed from '2G' to '2048', as '2G' is not
understood by autopkgtest-virt-qemu, while both autopkgtest-virt-qemu and
qemu understand the ram size as a plain number, and assume it's in megabytes.

This issue has been discovered as the Ubuntu Deskop upgrade from Disco to
Eoan takes more than 1GB of ram (qemu's default).