d/tests: replace broken `upgrade` test with simple `--help` call
The package had an autopkgtest that called the `upgrade` script.
That's not going to work as the `upgrade` script requires preparation
done via the `auto-upgrade-testing` command and that is based on a
configuration file (the test specification passed via --config).

In principle auto-upgrade-testing could itself be tested via an
autopkgtest, but that requires some non trivial work. For now let's
remove the broken autopkgtest and add a superficial `--help` call,
useful to spot major issues with the package (e.g. see d25cdf9d95).