Simplify pre/post scripts with only one folder and avoid duplicating
Before this patch, everything was handled as though we had two very
separate and distinct folders for pre_scripts and post_scripts. However,
in the profile configuration, there is only one `script_location`
variable, that was later passed to both handlers for pre/post scripts,
meaning we ended with two differents folders of the very same content.