Allow setting 'prompt' explicitely, for more control in upgrade profiles
This allows a profile to set `prompt: lts` or `prompt: normal`, so
that upgrades from LTS to LTS can be forced to take either the short
path directly to next LTS, or the long path through interim releases,
depending on what the profile should test.
If `prompt` is not set, then the old behavior has been kept as a
fallback. It tests first with `prompt=lts` if the target can be reached,
then changes back to `normal` if the release is not found.