feat: Check the kernel booted during upgrade is the same as the kernel booted post-upgrade
This check looks to catch errors, such as the ones we saw with
jammy->noble upgrades:
https://discourse.canonical.com/t/retrospective-on-jammy-noble-kernel-header-upgrade-issue/4263

Prior to upgrade, the upgrade script now stores the output of:
dpkg -l linux-*-$kernel

where $kernel=$(uname -r).

Then, post upgrade, we check the output of the dpkg -l... command again.

If it is the same - happy! If not, fail :'(

This should mean that we spot upgrade issues related to kernels and
kernel headers much quicker in the future.