Make sure '/tmp' is a tmpfs that gets cleaned
With `systemd` v256 and LP: #2019026, `/tmp` was made a `tmpfs` by
default. This led to LP: #2069834, and `autopkgtest` now has a mechanism
to prevent `/tmp` to be a tmpfs. Probably due to a bug that still
requires investigation, `/tmp` is not cleaned upon reboot, making our
canary to stay, and the upgrade to fail inappropriately.
This patch removes the `autopkgtest` created file that prevents `/tmp`
to be a `tmpfs`, making the upgrade green again.

As mentionned, `tmpfs` or not, `/tmp` should be cleaned, meaning further
investigation is needed there to get to the bottom of it.