properly reboot initial testbed when needed
Before running `do-release-upgrade`, a testbed needs to be fully up to
date, and potential kernel upgrades applied, meaning the testbed must be
rebooted on the latest installed kernel.

This was somehow running fine (although it's a mystery it worked) with
autopkgtest 5.38ubuntu1~22.04.1 on Jammy (the Jenkins runner), but got
broken after the backport of autopkgtest 5.47~22.04.1, which is now
exposing the correct behavior, breaking the upgrades on not up-to-date
testbeds.

This commit ensures we have the testbed fully upgraded and rebooted
before proceeding with `do-release-upgrade`.

LP: #2115731